Update update_code.pl
WIP

+++ vagrant_provision/pl/update_code.pl
@@ -5,17 +5,9 @@ use strict;
 use warnings;
 use autodie;
 use File::Basename qw(fileparse);
+use File::Copy qw(copy);
 use File::Path qw(remove_tree);
 use File::Spec::Functions qw(catfile rootdir);
-
-# rsync provides easy recursive copy, but is not part of Perl core libraries.
-sub exec_rsync {
-    my ($source, $dest) = @_;
-    if ((system "rsync -qa $source $dest") != 0) {
-        print "rsync $source: $!\n";
-        exit 1;
-    }
-}
 
 # Root required
 print "Root required.\n" and exit 1 if ($> != 0);
@@ -28,6 +20,13 @@ my @REPO_PATH = (rootdir(), "home", "vagrant", "github_phplw");
 my @HTML_PATH = (rootdir(), "var", "www", "html");
 my @CONFIG_PATH = (@REPO_PATH, "vagrant_provision", "config");
 
+my $VAGRANT_NAME = "vagrant";
+
+# HTML file ownership UID and GID
+my $HTML_UID = getpwnam("www-data");
+my $HTML_GID = getgrnam("www-data");
+my $HTML_PERMISSIONS = 0664;
+
 # Files
 my $CONFIG_FILE = "config.php";
 
@@ -36,22 +35,39 @@ my ($source, $dest, $files, @source_path, @dest_path);
 
 # ** -------------------------- END CONFIGURATION --------------------------- **
 
+# Help provide some error messaging if there is a copy error.
+sub copy_file {
+    my ($source, $dest) = @_;
+    print STDERR "copy $source: $!" and exit 1 if !(copy $source, $dest);
+    chown $HTML_UID, $HTML_GID, $dest;
+    chmod $HTML_PERMISSIONS, $dest;
+}
+
+if (defined $ARGV[0] && $ARGV[0] eq "composer") {
+    $dest = catfile(@HTML_PATH);
+    print STDERR "composer exited ", $? >> 8, "\n" and exit 1 if !(system "su -c \"composer -d$dest update\" $VAGRANT_NAME");
+
+    # CLI arg "composer" only updates composer code in working directory.
+    exit 0;
+}
+
 # Remove everything from HTML directory
 remove_tree(catfile(@HTML_PATH), {safe => 1, keep_root => 1});
 
 # Copy code to HTML directory
-foreach (qw(*.php *.html *.css vendor)) {
+foreach (qw(*.php *.html *.css *.json)) {
     $files = catfile(@REPO_PATH, $_);
     foreach $source (glob($files)) {
-        $dest = catfile(@HTML_PATH);
-        exec_rsync($source, $dest);
+        $files = fileparse($source);
+        $dest = catfile(@HTML_PATH, $files);
+        copy_file($source, $dest);
     }
 }
 
-#Copy dev config file
+# Copy dev config file
 $source = catfile(@CONFIG_PATH, $CONFIG_FILE);
 $dest = catfile(@HTML_PATH);
-exec_rsync($source, $dest);
+copy_file($source, $dest);
 
 # All done!
 exit 0;
